fix(db,cli,orm): PR-17 follow-ups — JSON (not JSONB) for raw_record; Decimal typing for category_confidence; reorder CLI to persist before categorize
- libs/db/alembic/versions/0001_fa_core.py: use sa.JSON for fa_transactions.raw_record and drop dialect import; keep column-level unique on fingerprint_sha256
- libs/db/src/db/models/finance.py: category_confidence → Mapped[Decimal | None] (Numeric(3,2))
- packages/financial_analysis/cli.py: when --persist, upsert transactions first (short txn), run categorization, then apply category updates in a second short txn

+++ libs/db/src/db/models/finance.py
@@ -79,7 +79,7 @@ class FaTransaction(Base):
         nullable=False,
         server_default=text("'unknown'"),
     )
-    category_confidence: Mapped[float | None] = mapped_column(Numeric(3, 2), nullable=True)
+    category_confidence: Mapped[Decimal | None] = mapped_column(Numeric(3, 2), nullable=True)
     categorized_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     is_deleted: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("false"))
     created_at: Mapped[datetime] = mapped_column(
